increased the smoothing factors
Though this looks like a lot of changes, most of it is debug and minor stuff.  The most important change here is inreasing the RTTSMOOTHFACTOR_UP/DOWN numbers each by a factor of 4.

This achieves my goal of slowing down the changes so that modest changes in latency won't flip/flop the LT so much.  My goal was, for an LT change of 2, the LT calculation would change by about 1.0  in the first LatencyVote and by about 0.5 then 0.25 in subsequent LatencyVotes.  While it's not an exact science, these numbers seem to do what I want.

On the downward side, a change from LT=4 to LT=2 should follow a similar pattern but should take about twice as long as the upward change.

The real test will be in the real world.

+++ src/net/CUDPComm.cpp
@@ -1873,7 +1873,7 @@ void CUDPComm::Reconfigure() {
 }
 
 long CUDPComm::GetMaxRoundTrip(short distribution) {
-    long maxTrip = 0;
+    float maxTrip = 0;
     CUDPConnection *conn;
 
     for (conn = connections; conn; conn = conn->next) {
